EZP-26057: move PermissionResolver implementation to Core/Permission
