+++ eZ/Publish/Core/MVC/Symfony/Cache/Tests/Http/EventListener/RelatedLocationsListenerTest.php
@@ -70,7 +70,7 @@ class RelatedLocationsListenerTest extends PHPUnit_Framework_TestCase
     private function getPermissionResolverMock()
     {
         return $this
-            ->getMockBuilder('\eZ\Publish\Core\Repository\PermissionResolver')
+            ->getMockBuilder('\eZ\Publish\Core\Repository\Permission\PermissionResolver')
             ->setMethods(null)
             ->setConstructorArgs(
                 [
