+++ eZ/Publish/Core/MVC/Symfony/Routing/Tests/UrlAliasGeneratorTest.php
@@ -423,7 +423,7 @@ class UrlAliasGeneratorTest extends PHPUnit_Framework_TestCase
     protected function getPermissionResolverMock()
     {
         return $this
-            ->getMockBuilder('\eZ\Publish\Core\Repository\PermissionResolver')
+            ->getMockBuilder('\eZ\Publish\Core\Repository\Permission\PermissionResolver')
             ->setMethods(null)
             ->setConstructorArgs(
                 [
