+++ eZ/Publish/Core/MVC/Symfony/Security/Tests/User/Identifier/RoleContextProviderTest.php
@@ -159,7 +159,7 @@ class RoleContextProviderTest extends PHPUnit_Framework_TestCase
     protected function getPermissionResolverMock()
     {
         return $this
-            ->getMockBuilder('\eZ\Publish\Core\Repository\PermissionResolver')
+            ->getMockBuilder('\eZ\Publish\Core\Repository\Permission\PermissionResolver')
             ->setMethods(null)
             ->setConstructorArgs(
                 [
